rename: test 폴더 example로 폴더명 변경
