+++ src/main/java/com/example/inkspire/example/TestController.java
@@ -1,9 +1,9 @@
-package com.example.inkspire.test;
+package com.example.inkspire.example;
 
 import com.example.inkspire.common.DataResponseDto;
 import com.example.inkspire.common.PageResponseDto;
 import com.example.inkspire.common.ResponseDto;
-import com.example.inkspire.test.model.TestDto;
+import com.example.inkspire.example.model.TestDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
