+++ src/main/java/com/example/inkspire/example/TestService.java
@@ -1,4 +1,4 @@
-package com.example.inkspire.test;
+package com.example.inkspire.example;
 
 import com.example.inkspire.common.DataResponseDto;
 import com.example.inkspire.common.PageInfoDto;
@@ -6,7 +6,7 @@ import com.example.inkspire.common.PageResponseDto;
 import com.example.inkspire.common.ResponseDto;
 import com.example.inkspire.common.errors.ErrorCode;
 import com.example.inkspire.common.errors.GeneralException;
-import com.example.inkspire.test.model.TestDto;
+import com.example.inkspire.example.model.TestDto;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
